Fixed issue with loading external entities (XML).
Fixed possible security issue with loading external entities from an XML
file. Starting with libxml2 version 2.9.2, the parser loaded and substituted
external entities (read from file) in an XML document. By doing so, unwanted
data could be leaked (such as the contents of the file "/etc/passwd") when
parsing an appropriately prepared XML file. This type of security issue is
known as XML External Entity (XXE) injection.

The issue was not present in libxml2 version 2.9.1 and since the DCMTK does
not explicitly pass the parser flag XML_PARSE_NOENT to the xmlReadFile()
function, which is actually meant for this purpose, this should also be
true for older versions of the libxml2 library (but has not been tested).

Thanks to Maria Samoylova <mashanedyak@gmail.com> for the report, the
comprehensive analysis and for proprosing a (temporary) solution.