Made building/installing of manpages configurable.
Building and installing of DCMTK's manpages is now an configurable option
(CMake) instead of being forcibly disabled on Windows. The defaults of this
option reflect the old behavior (OFF on Windows and ON everywhere else), but
Windows users are now free to enable it if they want it for whatever reason.