Made patching manpages platform independent.
The shell script "patchman.sh" and a call to "rm -f" was replaced with a CMake
script "patchman.cmake" such that this can now be run on Windows without
additional dependencies to a shell etc.
The existing script is kept for GNU Autconf that requires those dependencies
anyway and will be removed together with Autoconf support some time in the
future.