Enhanced performance of inserting sequence items.
Significantly enhanced the performance of inserting items at the end of a
DcmSequenceOfItems or DcmPixelSequence instance. These enhancements are
based on a smarter use of the underlying DcmList class.

Adding 99,999 items to a pixel sequence now only takes a fraction of a
second compared to a very large number of seconds (as it was before).

Also clarified that the item-related helper methods in the DcmItem class,
such as findAndGetSequenceItem(), cannot be used with an instance of the
DcmPixelItem class (since this class is not derived from DcmItem).