Fixed bug affecting multiple nested signatures.
Fixed bug that caused digital signatures embedded in a sequence item to be
included in the bitstream of another sequence (e.g. on main dataset level)
that signed the sequence containing this sequence item. In this case the
bitstream generated for the signature creation or verification was
not DICOM compliant.