Fixed possible issue with getFullOverlayData().
Fixed issue with getFullOverlayData() that could cause a crash (segmentation
fault) when processing an overlay plane with an origin other than (0,0).
Note that the dcm2pnm/dcmj2pnm/dcml2pnm command line tool does not use this
method and is, therefore, not affected by this issue.

This closes DCMTK Bug #911.

Thanks to Bartosz Bialoskorski <bartoszbialoskorski@gmail.com> for the report
and the sample DICOM file.