Improved timestamp verification code.
The timestamp verification code now performs additional consistency
checks (for example whether the timestamp was created before the start
of validity of the TSA certificate). Furthermore, a valid certified
timestamp can now extend the "lifetime" of a signature where the
signer certificate has expired, if the timestamp is still valid.