Fixed bug in the parsing of UN VR sequences.
Added command line option --convert-un to storescp and fixed a bug in the
parser that caused parse errors when receiving a large sequence element with
explicit length and UN value representation, while --convert-un is active.

Thanks to Jesper Bojesen <jbojesen@vitalimages.com> for the bug report and fix.

This closes DCMTK issue #888.