Introduced new internal VR "px" for pixel data.
Introduced new internal value representation "px" that can be used for pixel
data, both uncompressed (OB or OW) and compressed (pixel sequence). This new
type allows for treating other attributes than (7fe0,0010) in the same way as
the Pixel Data attribute, by simply specifying the tag in the data dictionary.

In a subsequent commit, this new feature will be used for the standard Pixel
Data attribute but also for private ones.