Fixed issue with order of include directories.
Fixed issue with order of include directories when working with CMake on
a system where the DCMTK libraries and header files are already installed
in the same directory as one of the external support libraries (e.g. zlib
or libiconv in "/usr/local").

Thanks to Steve Pieper <pieper@isomics.com> for the original report and
the in-depth analysis.