Updated and revised DIMSE Status Code definitions.
Updated and completely revised DIMSE Status Code definitions. Now, the
definitions and also the log output are consistent with the current
edition of the DICOM standard. Some Service Class specific DIMSE-N
Status Codes that are not needed for the DCMTK are still missing and
will be added in a future commit.

This closes DCMTK Conformance #803 and partly closes Conformance #598.