Fixed segmentation fault in OFStandard::atof()
Fixed a segmentation fault in OFStandard::atof() due to an out-of-bounds
access to an internal array in cases where the exponent overflows to
negative during the parsing process. The implementation is now consistent
with strtod(3): in these cases, plus or minus HUGE_VAL is returned if
the exponent is positive, and zero if the exponent is negative. In
both cases, the success flag is set to false.

Thanks to Maria Nedyak <mashanedyak@gmail.com> for the bug report.

This closes DCMTK issue #932.