Generalized support for Key Object Documents (KO).
Added more general support for SR IODs that contain the Key Object Document
Modules (on Series and Document level). This is needed since the Rendition
Selection Document IOD also contains these modules.

This is only a preparatory step for implementing support for this SR IOD
introduced with Supplement 202 (Real-Time Video).