Added support for new SR IOD from Supplement 202.
Added full support for the new Rendition Selection Document IOD introduced
with Supplement 202. Please note that the associated Rendition Selection
Document Real-Time Communication SOP Class is not a Storage SOP Class, so
instances of this class are not meant to be stored as files or transferred
over the network using the Storage Service Class.

Closes DCMTK Feature #926.