Output error message on missing sequence element.
Output error message on missing Verifying Observer Sequence element when
checking a DICOM file of the type "SR DOCUMENT" to be added to a DICOMDIR.
Previously, the file was rejected (e.g. by dcmgpdir or dcmmkdir) but the
reason was not given.

Thanks to DCMTK forum user "ruben.cruz" for the report and sample file.