New default to always re-create meta header.
The default behavior of DCMTK when writing DICOM files was to re-use
existing meta header information. This can lead to problems if essential
information in the dataset (like SOP Instance UID) changes without
explicitly enforcing those changes in the meta header, too.

By always re-creating the complete meta header it is ensured that all
its attributes are up-to-date and complete.