Fixed issue with invalid length of user item.
Fixed issue with an invalid value in the item-length field of the Maximum
Length Sub-Item Structure (A-ASSOCIATE-RQ). In case of a malformed
association request, the non-existent test on the item-length value could
result in a crash (segmentation fault).
Now, the value in the item-length field of this sub-item type (51h) is
checked (similar to the check for the other sub-items).

Thanks to Maria Nedyak <mashanedyak@gmail.com> for the bug report as well
as the test data and script that allowed for reproducing this issue.