Fixed JSON InlineBinary pixel data encoding.
Implemented a new method DcmPixelData::writeJson() that properly handles
the InlineBinary JSON encoding of DICOM pixel data, which is not supported
in the DICOM JSON model for encapsulated (compressed) images. Instead of
silently writing an empty pixel data element, we now report an error.

Together with various previous commits, this closes DCMTK issue #881.