Fixed issue with placeholders (--exec-on-eostudy).
Fixed issue with placeholders not being populated for the --exec-on-eostudy
option when receiving the first batch of images.

This issue was apparently introduced with commit 26441a226, when fixing
another issue regarding the use of --exec-on-eostudy together with --fork.

Thanks to Andreas Keizers <an-kei@web.de> for reporting this issue.

This closes DCMTK Bug #880.