Fixed wrong calculation of sigmoid VOI function.
Fixed possibly wrong output of sigmoid VOI LUT function, e.g. when
processing an image with a photometric interpretation of MONOCHROME1
(i.e. inverse). The original formula in the DICOM standard was incorrect
and has been fixed only recently with CP-1880.
Please note that the issue was only present when no presentation LUT or
display calibration was enabled (case #6 and #8 of the monochrome "VOI
SIGMOID" rendering algorithm).

Thanks to Robert Mulcahey <rob@asteris.com> for the report and suggested fix.