Adjusted the number of digits printed for FL and FD.
Adjusted the number of digits printed for FL elements to FLT_DECIMAL_DIG
(9) instead of 8, to make sure that the printed values can be converted
back to FL without loss. Fixed comment for FD, where the values was
already correct.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for pointing
out the issue.