Introduced a CMake option for static linking.
Introduced DCMTK_LINK_STATIC as an option for controlling how DCMTK's tools
and libraries are linked, affecting linker options and third party libraries.

These changes originate from our (so far not publicly available) release build
scripts. I noticed the way this is implemented is not that elegant while
creating this commit, but this way has the advantage that it has been verified
to work by the last couple of DCMTK releases. Most notably, the feature
removes flags from the CMake variables for dynamic linking to produce
statically linked binaries. We should better find a way to tell CMake it
should use the respective CMake variables for static linking instead in the
first place.

For now, the main goal of this commit is to make this feature publicly
available while keeping the build scripts consistent with it. Enhancing the
way the feature is implemented should (and hopefully will) come later.