Fixed extra padding created for some segmentations.
The problem showed up if the number of pixels per frame is not divisable by 8
and the the number of pixels per object (i.e. all frames) is. In that
case, one byte extra padding is added at the end, which (if Pixel Data
length is odd then), leads to another extra byte since element length
must always be even in DICOM.