Don't mask error code of association negotiation failure.
So far DCMTK always returned DUL_REQUESTASSOCIATIONFAILED when an attempt to
open a network connection by a client failed, independent of the reason.
This masked for example the specific error codes issued by the TLS module
which explain the reason for a TLS handshake failure. Changed the network
module to directly return this lower-level error code.