Fixed issue with findAndGetUint16Array and VR=AT.
Fixed issue with wrong number of 16-bit values being returned for Attribute
Tag (AT) values in the "count" parameter of the findAndGetUint16Array()
method. The return value was half as large as expected.

Also added a new regression test for this purpose.

Thanks to Jason LAMBERT <lambert@dosisoft.fr> for the report.