Changed default TLS profile to BCP 195 ND.
Changed the default security profile in module dcmtls from the BCP 195 TLS
profile to the Non-downgrading BCP 195 TLS Profile, since the BCP 195 TLS
profile is not compliant anymore to BCP 195 as amended by RFC 8996 in
March 2021. TLS protocol versions 1.0 and 1.1 are now explicitly forbidden
in BCP 195.

This closes DCMTK issue #973.