Introduced OFdeprecated and OFdeprecated_msg.
Introduced two new macros for marking functions as deprecated, with and
without an optional custom message.
The underlying compiler features should in theory also work on various kinds
of type and variable definitions, however, this was not tested and is also not
reflected in the introduced configuration tests. Therefore, the documentation
describes this feature only in relation with functions for now, which may be
extended as required after more testing.

This closes DCMTK issue #887.