Add DCMTK_NO_TRY_RUN for cross compiling
When doing cross compilation you can't always run the code on the target
device. So instead of doing a TRY_RUN, we are able to disable it and set
the values directly from the cmake command line.

Though the user also need to provide an arith.h file.

Example when pwd is located inside a build folder created in the root of
the git archive, and running on Linux on x86_64 and target also is
x86_64. Please look into GenerateDCMTKConfigure.cmake for further info.

cmake -DCMAKE_SYSTEM_NAME:STRING=Linux \
-DDCMTK_NO_TRY_RUN:BOOL=1 \
-DC_CHAR_UNSIGNED:BOOL=0 \
-DDCMTK_ICONV_FLAGS_ANALYZED:BOOL=TRUE \
-DDCMTK_FIXED_ICONV_CONVERSION_FLAGS:STRING=AbortTranscodingOnIllegalSequence \
-DDCMTK_STDLIBC_ICONV_HAS_DEFAULT_ENCODING:BOOL=TRUE \
..

Setting the CMAKE_SYSTEM_NAME sets the CMAKE_CROSSCOMPILING variable.

Signed-off-by: Claus Stovgaard <clst@ambu.com>