Removed useless typecasts from inline method.
Removed useless typecasts reported by gcc with -Wuseless-cast. There are
numerous other warnings like this in the DCMTK that are reported when the
additional warning option is set, but these two could leak into user code
since they are used in an inline method.

Thanks to GitHub user "malaterre" for the report and suggested fix.