Fixed issue with findAndGetUint16Array() for OB.
Fixed issue with findAndGetUint16Array() when accessing pixel data that
is stored with a VR of "OB" (Other Byte). In this case, this method
returned an incorrect value for the optional "count" parameter. Of
course, the method findAndGetUint8Array() should be used for "OB" data,
but the returned number of elements in the array should be correct in
any case.

Now, all findAndGetXXXArray() methods again use the data element length
for determining the value of the "count" parameter, which requires no
handling of special cases.

Thanks to forum user "st80rules" for the report.