Fixed Media Storage SOP Class UID in DICOMDIR.
Fixed wrong value for Media Storage SOP Class UID (0002,0002) that was
added to the File Meta Information Header of a DICOMDIR file. Instead
of 1.2.840.10008.1.3.10 (Media Storage Directory Storage SOP Class) the
private DCMTK-specific SOP Class UID 1.2.276.0.7230010.3.1.0.1 was used.
This was a result of a change of the default parameter of the method
DcmFileFormat::validateMetaInfo() introduced with commit d31f92867e.

Thanks to Gilles Mevel <gilles.mevel@groupe-nehs.com> for the report and
suggested fix.