Corrected handling of character set for HG images.
The dcmpsprt tool reads an image (and optionally a presentation state)
and prepares a print job that is stored as a set of SP and HG files,
which can be spooled to a print server using dcmprscu. The HG image
contained some patient attributes (such as the PatientName) of the
original image, but did not proporly copy SpecificCharacterSet.

Thanks to Daniel Calvo <dcalvo@fitec.org.br> for the bug report and fix.