Added CMake option DCMTK_PORTABLE_LINUX_BINARIES.
Added CMake option DCMTK_PORTABLE_LINUX_BINARIES that is intended to help
with the generation of binary executables that run on different distributions.
When this option is enabled, only static versions of the third party libraries
are detected and used, and the linker is run with the options -static-libgcc
-static-libstdc++, which cause a static version of libgcc and libstdc++ to be
used. Note that glibc is still dynamically loaded, because a statically linked
glibc actually causes portability issues due to glibc components loaded from
shared objects at runtime.

The new option does not yet handle the link dependencies of the external
libraries (such as libxml2 requiring zlib) properly. In order to handle these,
one would have to identify the dependencies using pkg-config and then search
for static versions of all dependencies, possibly recursively.