Replaced C_CHAR_UNSIGNED test by a compile-only one.
Replaced the test for the signedness of the char type, which so far
needed to execute a test program with TRY_RUN, with test code that only
needs to compile, but not run, which makes things easier in a cross-compilation
setting.

Thanks to Helmut Grohne <helmut@subdivi.de> for the suggestion and patch.
This closes Github pull request #44.