Initial commit of BSD citrus libiconv.
Initial commit of the citrus libiconv implementation, which is licensed under
the BSD license and part of the FreeBSD and NetBSD operating systems.
This commit retains the original files as taken from the FreeBSD git
with only the minimum set of changes required to make the library compile
(but not yet work) on Linux. All changes are marked with
BEGIN/END DCMTK modifications comments.