Fix some binary segmentation concatenations.
When producing binary concatenations of segmentations objects using
a configuration where the number of bytes per concatenation instance is
not dividable by 8 (i.e. where in such an instance numFrames * rows *
cols % 8 != 0), not all bits (pixels) have been copied correctly from
the source image into the concatenation instance.

The problem has been fixed by adapting number of frames in the
concatenation instance so that numFrames * rows * cols % 8 = 0. If the
user configures numFrames (per instance) manually, the value is adapted
automatically as well (and an information is printed to the logger).

Also, when loading 1 bit / pixel segmentation concatenations, there was
a bug with clearing certain bits due to the fact that bit-shifting a
number to the right does not automatically introduce 0 but instead
introduceds 0 or 1 based on the signedness of the number's variable.