Improved detection of additional depencencies for statically linking.
Improved detection of additional library depencencies for statically
linking the third-party libraries used in the DCMTK. The problem is caused by
third party libraries having depencencies themselves. For example libxml2
requires zlib among others. This gets even more complicated when on one
operating system for example libtiff is compiled with libdeflate an on
another system it is not. This means on one system there are linker errors
when libdeflate is not included and on another system there are errors when
it is included.
The best solution would possibly be to just compile all third-party libraries
ourself and use these versions for statically linking. In the meantime this
solution can somewhat handle these differences.