Forbid insertion of encapsulated Pixel Data.
dcmodify's options -mf and -if can be used to insert the content of a
file into an element's value. This also works for uncompressed data in
the Pixel Data element.

If a user tried to replace or insert encapsulated pixel data, the new
value was not inserted at all, however, dcmodify reported success.

Now, the insertion/modification of values for encapsulated Pixel Data is
rejected by dcmodify and a related error is reported to the user.

Also, trying to insert value directly from the command line into the
encapsulated Pixel Data element (via -m / -i options) now produces a
better error message.

Thanks to forum user "upnorth173" for the report.