Fixed RescaleType when lossy compressing CT.
When a CT image is compressed with lossy JPEG and the attributes
RescaleSlope/RescaleIntercept are present, the RescaleType attribute
was set to "HU" (Hounsfield units), which was correct at the time
when the dcmjpeg module was created, but is not anymore, since the
CT image module nowadays permits other Defined Terms. Therefore, the
original value of RescaleType is now retained for CT images.