Added check on missing Unique Key (strict mode).
Report a more appropriate error, both to the logger and as a DIMSE Status
Code (A900 instead of C000), if the Unique Key is missing for the higher
level(s) of a C-FIND request.

This check is only performed in "strict mode", i.e. when dcmqrscp is called
with option --check-find (-XF).

Thanks to Abdallah Maouche <abdallahmaouche98@gmail.com> for the report and
suggested fix. See also PR #53 at https://github.com/DCMTK/dcmtk/pull/53.