Don't add Derivation Description to Icon Images.
When compressing a DICOM image with an embedded icon image with the JPEG
true lossless encoder, which is dcmcjpeg's default compression option, a
Derivation Description (0008,2111) was added to the Icon Image Sequence
(0088,0200), even though it should only be added to the top-level dataset.
This issue has been fixed now.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the
original report.