Emulate libiconv behavior when using oficonv.
By default, GNU libiconv initializes a conversion descriptor with
transliteration disabled (i.e. a conversion will stop with an error
if a character cannot be mapped to an equivalent counterpart in the
target character set). In oficonv this mode is not available but a
very similar behavior can be activated by configuring the
ICONV_SET_ILSEQ_INVALID mode. This is now done by default when using
oficonv.