Added possibility to embed oficonv data in library.
Added code that allows for the generation of an oficonv library
where all data files are statically included and do not have
to be loaded from the file system. This is useful primarily when
oficonv is compiled into a shared library. This mode is activated
when compiling with the macro DCMTK_ENABLE_BUILTIN_OFICONV_DATA.
Note that DCMICONVPATH is still evaluated when this mode is active
and can be used to override the built-in datasets at runtime.