Fixed issue when using option +Ca and +U8.
Fixed an issue, or rather an unexpected behavior, when using option +Ca
(--charset-assume) together with option +U8 (--convert-to-utf8).

Now, when calling the xml2dcm tool for a DICOM file that is missing the
SpecificCharacterSet attribute although it is needed (because extended
characters are used), option +Ca is processed first and then +U8. This
allows for specifying the correct value for the missing attribute (e.g.
"ISO_IR 100") before converting the dataset and also the XML output to
UTF-8 encoding.

This partly closes DCMTK Feature #1024.