Added function ASC_setParentProcessMode().
Added a new function ASC_setParentProcessMode() that must be called by
the parent process after a fork() operation if a network connection has
already been accepted and is further handled by the child process.
In the case of TLS connections, this will prevent ASC_dropAssociation()
in the parent process from issuing a TLS close_notify shutdown alert
to the peer.