Completed support for CP-2000 (dcmsr/cmr).
Completed support for CP-2000 by adding support for the SR Template
parameters $ImagePurpose or $Purpose to the implementation of TID 1501,
300 and 320. This way, codes from CID 7551 can be passed as "types" e.g.
to the method TID300_Measurement<>::addImage() or addSpatialCoordinates().

Please note that an instance of the DSRCodedEntryValue class can still
be passed to these methods, but a CODE_DCM_xxx define has to be converted
explicitly (see test case "dcmsr_TID1501_MeasurementGroup" for an example).

This closes DCMTK Conformance #950.