Fix DCMTK_ENABLE_STL=TRUE not handled correctly
DCMTK_ENABLE_STL_${FEATURE} are "inferable" variables that can have
one of the values "ON" (truthy), "OFF" (falsy), or "INFERRED". If the
value is "INFERRED" it means the variable will be set to the value of
DCMTK_ENABLE_STL.

After the special case "INFERRED" has been handled
DCMTK_ENABLE_STL_${FEATURE} shall be treated like a normal CMake
boolean expression. It shall therefore not be compared for equality to
"ON", rather the truthiness shall follow the normal rules of CMake
variables, which means that a variable is true unless it has the value
0, OFF, NO, FALSE, N, IGNORE, NOTFOUND, is the empty string, or ends
in the suffix -NOTFOUND.