Fixed --version output when STL support enabled.
Fixed the --version output when the DCMTK was compiled with STL support
(C++ Standard Template Library) enabled. Now, the output is "stl" if all
HAVE_STL_xxx macros are defined and "(stl)" if at least one of the
HAVE_STL_xxx macros is defined.

Before, "stl" was never printed, even if the DCMTK was compiled with STL
support enabled, because the outdated macro HAVE_STL was checked.