Added classes for handling IPC message queues.
Added classes OFIPCMessageQueueServer and OFIPCMessageQueueClient, which
implement the server and client side of a message queue for inter-process
communication. A server process can create a message queue and receive
text messages, and a client process can open the queue and send messages.
Depending on the platform, the queues are implemented as Windows Mailslot,
Posix message queue or System V message queue.