Make DcmSCU TCP connection timeout configurable per instance.
This change makes it possible to configure the DcmSCU TCP connection timeout per instance, instead of using the global dcmConnectionTimeout parameter.

In a multi-service executable, that for example supports QR, Worklist, and Storage, storage may be executed in parallel with Worklist and QR. This is a typical pattern on some DICOM modalities, where the user can do QR or Worklist queries while images are being sent in the background.

It is then beneficial to be able to configure different TCP connect timeouts for each individual service. A typical example could be a multi-site institution where the storage server is far away and requires a long TCP connection timeout, whereas the Worklist server could be on site and require a short TCP connection timeout.

This becomes particularly important for mobile devices, where the device may have access to a storage server, temporarily not have access to the worklist server. We would then like to have a short worklist timeout to quickly fall back to an offline worklist.