Make DcmSCU TCP connection timeout configurable per instance.
This change makes it possible to configure the DcmSCU TCP connection timeout per instance, instead of using the global dcmConnectionTimeout parameter.

In a multi-service executable, that for example supports QR, Worklist, and Storage, storage may be executed in parallel with Worklist and QR. This is a typical pattern on some DICOM modalities, where the user can do QR or Worklist queries while images are being sent in the background.

It is then beneficial to be able to configure different TCP connect timeouts for each individual service. A typical example could be a multi-site institution where the storage server is far away and requires a long TCP connection timeout, whereas the Worklist server could be on site and require a short TCP connection timeout.

This becomes particularly important for mobile devices, where the device may have access to a storage server, temporarily not have access to the worklist server. We would then like to have a short worklist timeout to quickly fall back to an offline worklist.

Fix issue with ambiguity of the value -1 in TCP connection timeout.

Instead of using -1 to indicate fallback to the timeout value specified by dcmConnectionTimeout, we can just remove the fallback and instead call dcmConnectionTimeout.get() directly where we want to specify the timeout value.

This way we avoid confusion related to what the value -1 means. Now -1 means infinite timeout.

Add TCP connection timeout argument to ASC_createAssociationParameters in dcmpstat and dcmqrdb projects.

I had previously only compiled with a subset of the DCMTK modules in my solution.

Add a couple of tests that verifies the new behavior of DcmSCU::setConnectionTimeout.

These tests helps documenting the behavior and prevents regressions in the handling of connection timeouts.

Fix the type of DcmSCU::m_tcpConnectTimeout which should be a Sint32.

A value of -1 is used to indicate inifinite timeout, so we should use a signed type.