First sketch for mechanism to cancel DcmSCU negotiateAssociation before connectionTimeout elapses.
This helps quickly terminating attempts to establish an association for example during applicaiton shutdown.

The problem is that if a server is unavailable, the negotiateAssociation function call will block for the duration of the connectionTimeout. In many cases this timeout may be long, 30-60 seconds to ensure that we can robustly establish associations against slow servers. This is a long time for a user to wait for a device to shut down.

In some use cases, for example for portable imaging devices, the ability to quickly shut down is important to reduce the risk of the user pulling the power cable before the system is completely shut down. Such devices typically have a high risk that servers are temporarily unavailable, for example if they are serving multiple purposes and have to be moved from lab to lab.

The design idea is that the client software can create a socketpair that is used as a communication channel for signalling cancellation. While an association is being established, the client software can send a message on the socket pair from a different thread, which will wake the select in dulfsm.cc requestAssociationTCP function. The error handling will then follow the same path as if the connectionTimeout would be reached.