Re-added old ASC_createAssociationParameters().
Starting with commit d468490a8, function ASC_createAssociationParameters()
expects the value for the TCP connection timeout as a third parameter.
For reasons of backward compatibilty, the old function (with only two
parameters) that uses the global setting for the TCP connection timeout
is re-introduced.

Please note, however, that this old function is deprecated and will be
(finally) removed in the future. See compiler warnings for details.