Added dcm2json command line options.
Added new command line options to dcm2json that allow the user to control the
policy for the conversion of IS/DS values to JSON. Users can either continue
to use the automatic conversion where values are converted to number if
possible and to string otherwise, or they can enforce conversion to string
or to number. In the latter case, an IS/DS value that cannot be mapped to
a JSON number will lead to an error message and the conversion will be
aborted.

This closes DCMTK feature request #988.