Updated data dictionary for DICOM 2022c.
Updated data dictionary for the latest edition of the DICOM standard, which
has been released only recently.

Since one of the DICOM attributes was "retired", the name of the associated
DcmTagKey constant also changed. This required minor changes to the source
code of the dcmrt module (automatically generated source files).