Added cleanup signal handler to ofipc code.
Added signal handler code to ofipc that can be used to ensure that message
queues are properly closed and deleted even when the application is terminated
abnormally through SIGINT (CTRL-C) or SIGTERM (kill). Otherwise, some types
of message queues such as Posix or System V will persist after the termination
of the process and count against the (relatively small) limit of message
queues permitted per user.