Make DcmSCU::nextMessageID protected, not private.
DcmSCU provides several virtual functions where the intent is to enable
users to override these virtual functions. For example,
DcmSCU::sendMOVERequest advertises that it can be overridden in the doc
string.

To be able to override these functions in a meaningful way, we need access
to the nextMessageID function from derived classes.

A use case for overriding DcmSCU::sendMOVERequest could be to enable
concurrent C-MOVE requests. In such cases, we need a way to get the
message ID out of the DcmSCU to be able to correlate it with the Move
Originator Message ID from a client side STORE SCP.

Also, improved wording of DcmSCU::sendMOVERequest doc string to
help readability.

Thanks to GitHub user "jogerh" (Joger Hansegard) for the proposed patch.