Refactor network init into separate method.
Move code related to network initialization within
DcmBaseSCPPool::listen to the separate function and make it virtual.

This way it is possible for the clients to implement a notification about
the results of the network initialization part. It is needed to be able to
get feedback for the cases where, for example, the selected port is already
taken.

Thanks to GitHub user "doskachok" (Vasyl Horbatenko) for the proposed patch.