Added support for official UID keywords.
Added support for the official DICOM UID keywords that were originally
introduced with CP-1995. These keywords are defined in addition to the
DCMTK-specific UID names that have been used for a long time and will
still be used for the foreseeable future.

During the implementation, a few minor issues were detected:

- The official name of the UID "1.2.840.10008.5.1.4.1.1.88.70" is
  "Implantation Plan SR Storage" and not "Implantation Plan SR
  Document Storage". Also see CP-2240 (working draft).

- The official name of the UID "1.2.840.10008.1.4.3.2" is
  "Standard Robotic-Arm Coordinate System Frame of Reference" and
  not "Standard Robotic Coordinate System Frame of Reference".

- The definition of the UID "1.2.840.10008.5.1.4.38.4" was missing in
  the list of UIDs known to the DCMTK, i.e. the Hanging Protocol
  Information Model - GET SOP Class.

These issues were fixed in a backward compatible manner (if possible).

It is planned to support these official keywords in various DCMTK tools
with future commits, e.g. for selecting the preferred Transfer Syntax
on the command line or as an alternative to the DCMTK-specific name in
a textual dump.

This closes DCMTK Conformance #935.