Introduced DIMSE backward compatibility option.
Introduced a new macro DCMTK_ENABLE_ACR_NEMA_DATASET_PRESENT_COMPATIBILITY
that will cause DIMSE_DATASET_PRESENT to be defined as 0x0102, a value
that is required for backward compatibility with ACR/NEMA. This is not
enabled by default because it triggers a bug in previous versions of some
DCMTK tools, see previous commit and DCMTK issue #1045.

This closes DCMTK issue #238.