Fixed frame size calculation for YBR_FULL_422.
Added additional parameter to DcmElement::getUncompressedFrameSize()
that determines whether pixel data is compressed or uncompressed.
This is important for calculating the frame size for YBR_FULL_422 images,
which can exist both in compressed and uncompressed encoding.
Since our decoders convert such images to YBR_FULL during decompression,
the frame size is different. Furthermore, many uncompressed images claiming
to use YBR_FULL_422 photometric interpretation are in fact encoded in
YBR_FULL. We check the size of the pixel data element to account
for this issue.