Updated data dictionary for DICOM 2022e.
Updated data dictionary for the latest edition of the DICOM standard, which
has been released only recently.

Since two of the DICOM attributes were "retired", the name of the associated
DcmTagKey constants also changed. This required minor changes to the source
code of the dcmrt module (both automatically and manually generated source
files).