Presuppose C++ compiler support for C++ casts.
Currently DCMTK offers the macros OFstatic_cast(), OFconst_cast(),
OFdynamic_cast() and OFreinterpret_cast() in ofcast.h that map related
calls to C-style tasks if the corresponding C++ casts are not supported
by the compiler.

The availability of casts is checked on build configuration time by
setting HAVE_XXX_CAST in osconfig.h. At least this is done when compiling
with autoconf. However, since the introduction of CMake into DCMTK, CMake
scripts instead hard-code all HAVE_XXX_CAST macros to 1 (i.e.
available).

So far we never received reports of (CMake) DCMTK users that there are
still systems with compilers that don't offer all C++ casts by default,
i.e. which would rely on the DCMTK mapping to old-style tasks.

That is why we decided to now print an error if an (autoconf) user
actually uses a compiler that needs the mapping to old-style cast. This
is mainly to find out whether ther are still compilers in use that
require the mapping to old-style casts. The next step would be to remove
this mapping completely once we don't hear any complaints.