Print syntactically valid Json for compressed images.
The dcm2json tool does not support the conversion of DICOM datasets with
encapsulated pixel data to Json with InlineBinary encoding. The underlying
method DcmItem::writeJson() returns EC_CannotWriteJsonInlineBinary.
This modification makes sure that in this case the Json dataset produced
is still at least syntactically valid Json.

Thanks to Juha Reunanen for the report and pull request.
This closes GitHub pull request #77.