Fixed issue with DcmCodec::updateImageType().
When compressing a DICOM image, e.g. with JPEG Baseline, the JPEG
encoder calls DcmCodec::updateImageType(), which used to add the
Image Type (0008,0008) Attribute with a possibly invalid value.
For example, compressing a DICOM Secondary Capture image that does
not contain the Image Type (0008,0008) Attribute resulted in an
invalid value of "DERIVED" for the Image Type (0008,0008) Attribute.
This value is invalid, because the VM is "2-n" according to DICOM
PS3.6, i.e. at least two values should be present.

This and other related issues have been fixed by completely rewriting
the updateImageType() method. Also, the API documentation now clearly
states which cases are handled in what way. And, the method reports
details on what is done or what failed to the debug/error logger.

This closes DCMTK Bug #1056.